refactor(shims): split NextRequest primitives from next server
The public next/server shim currently owns both the exported request and response classes and the lower-level request primitives they depend on. That makes the file harder to review because NextURL, RequestCookies, request sealing, URL validation, and NextFetchEvent sit alongside NextRequest and NextResponse behavior.

Extract those request primitives into an internal module and re-export them from next/server so the public API remains unchanged. Route handlers continue to construct real NextRequest instances; this commit does not change the request wrapper behavior.
